+++ dev-viewer/src/js/index.js
@@ -2,16 +2,15 @@ import 'core-js/stable/promise';
 import './ui';
 import './storage';
 
-import {LoggerFactory, LogLevel} from "@srag/pdf-wrap";
-
-import {PDFjsDocumentService, setAssetsSrc} from "@srag/pdf-wrap";
+import {PDFjsDocumentService, setAssetsSrc, LoggerFactory, LogLevel} from "@srag/pdf-wrap";
 import {URI} from "@srag/pdf-wrap";
 
 import {SelectionButton} from "./tools/selection";
 import {FormsService} from "./tools/forms";
-import {ClearButton, HighlightButton, HighlightService} from "./tools/highlight";
+import {HighlightService} from "./tools/highlight";
 import {EraserButton, PenButton} from "./tools/freehand";
 import {SidebarManager} from "./sidebar";
+import {ScalePreset} from "@srag/pdf-wrap";
 
 
 LoggerFactory.configure({
@@ -39,6 +38,9 @@ export async function loadPDF() {
     });
 
     pdfDocument.scaleTo("page-fit");
+    pdfDocument.zoom.config.minScale = 0.1;
+    pdfDocument.zoom.gesture.doubleTap.snap.enabled = true;
+    pdfDocument.zoom.gesture.doubleTap.snap.scaleTo = "page-height";
 
     new HighlightService(pdfDocument.highlighting);
     new PenButton(pdfDocument.toolbox.freehand);