+++ dev-viewer/src/js/ui.js
@@ -6,7 +6,7 @@ import 'bootstrap-colorpicker/dist/js/bootstrap-colorpicker.min'
 //CSS
 import '../css/normalize.css'
 //import '../css/skeleton.css'
-import '../../build/www/assets/libs/pdf-wrap/pdf-wrap.css'
+import '../../../pdf-wrap/assets/pdf-wrap.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-colorpicker/dist/css/bootstrap-colorpicker.min.css'
 import '../css/dev-viewer.css'