+++ pdf-wrap/.mocharc.cjs
@@ -1,0 +1,10 @@
+const CONFIG = {
+    spec: "./test/**/*.spec.ts",
+    extension: ["ts"],
+    package: "./package.json",
+    require: [
+        "ts-node/register/transpile-only",
+    ]
+};
+
+module.exports = CONFIG;