+++ pdf-wrap/.mocharc.js
@@ -1,3 +1,0 @@
-module.exports = {
-    spec: "build/javascript/test/**/*.spec.js"
-};