+++ pdf-wrap/src/api/document/index.ts
@@ -1,2 +1,3 @@
+export * from "./document.scale";
 export * from "./document.info";
 export * from "./pdf.document";