+++ pdf-wrap/src/api/zoom/interaction.ts
@@ -2,7 +2,18 @@
  * A generic zoom interaction, which is triggered by a user
  * which interacts with the pdf document.
  */
+import { Observable } from "rxjs";
+import { ScalePreset } from "../document";
+import { StateChangeEvent } from "../event";
+
 export interface ZoomingInteraction {
+
+    /**
+     * State change events are emitted when the interaction changes between the enabled and disabled.
+     *
+     * @see StateChangeEvent
+     */
+    readonly stateChange: Observable<StateChangeEvent>;
 
     /**
      * Set the current state of the zoom interaction.
@@ -18,17 +29,9 @@ export interface ZoomingInteraction {
 }
 
 /**
- * A touch double tap interaction which zooms in and out of a pdf document.
- *
- * Behaviour:
- * If the user double taps the document, the interaction will increase the scale of the document.
- * After a second double tap the document scale will be restored.
- *
- * This interaction will not restore the original scale value if the user manipulates the value by other means (pinch etc.).
- *
- * Zooms page by 20%.
+ * Base interface for all double tap interactions.
  */
-export interface DoubleTapZoomingInteraction extends ZoomingInteraction {
+export interface DoubleTapInteraction extends ZoomingInteraction {
     /**
      * The maximum time delta in ms between two taps,
      * which will trigger this interaction.
@@ -38,6 +41,21 @@ export interface DoubleTapZoomingInteraction extends ZoomingInteraction {
      * Defaults to 500ms.
      */
     maxDoubleTapDelta: number;
+}
+
+/**
+ * A touch double tap interaction which zooms in and out of a pdf document.
+ * The double tap zoom mode is exclusive and will deactivate other double tap zoom gestures.
+ *
+ * Behaviour:
+ * If the user double taps the document, the interaction will increase the scale of the document.
+ * After a second double tap the document scale will be restored.
+ *
+ * This interaction will not restore the original scale value if the user manipulates the value by other means (pinch etc.).
+ *
+ * Zooms page by 20%.
+ */
+export interface DoubleTapZoomingInteraction extends DoubleTapInteraction {
 
     /**
      * The zoom factor of the double touch gesture.
@@ -50,4 +68,57 @@ export interface DoubleTapZoomingInteraction extends ZoomingInteraction {
      * Invalid values are ignored.
      */
     zoomFactor: number;
+}
+
+/**
+ * A touch double tap interaction which snaps the document to a predefined value for example page height.
+ * The double tap snap mode is exclusive and will deactivate other double tap zoom gestures.
+ *
+ * Behaviour:
+ * If the user double taps the document, the interaction will zoom the page to a predefined value like page height.
+ */
+export interface DoubleTapSnapInteraction extends DoubleTapInteraction {
+
+    /**
+     * The preset which is used to scale the document if the gesture is
+     * detected.
+     *
+     * Valid values:
+     * - "page-width"   -> scales to page width
+     * - "page-height"  -> scales to page height
+     * - "page-fit"     -> tries to fit the page into the viewport
+     * - "page-actual"  -> scales document to its actual size
+     * - "auto"         -> automatically scales the document
+     *
+     * @throws TypeError - Throws when an invalid value is passed to the property.
+     */
+    scaleTo: ScalePreset;
+}
+
+/**
+ * Double tap interactions collection.
+ */
+export interface DoubleTabInteractions {
+    /**
+     * @see DoubleTapZoomingInteraction
+     */
+    readonly zoom: DoubleTapZoomingInteraction;
+    /**
+     * @see DoubleTapSnapInteraction
+     */
+    readonly snap: DoubleTapSnapInteraction;
+}
+
+/**
+ * Zoom interactions.
+ */
+export interface ZoomingInteractions {
+    /**
+     * Pinch zoom interaction settings.
+     */
+    readonly pinch: ZoomingInteraction;
+    /**
+     * Double tap interaction settings.
+     */
+    readonly doubleTap: DoubleTabInteractions;
 }